Extract the array pattern arithmetic from the connection expander
Which elements of a source array and a destination array a Connection_Pattern
value, or the default pattern, pairs up is arithmetic over array sizes. It says
nothing about connections. A flow specification between feature arrays reads the
same property with the same meaning, so it needs the same walk.

Move the pattern literals and the dimension walk out of ConnectionArrayExpander
into ArrayPatternExpansion, which hands every paired-up combination of indices
to a Pairing callback and reports through a Subject that names what is being
expanded. ConnectionArrayExpander keeps the decision of what a pair of indices
becomes, and supplies "connection" as the subject kind, so its four reports read
exactly as they did.

No behavior change: the 96 tests over connection arrays, patterns and connection
sets pass unchanged.
