+++ core/org.osate.aadl2.instantiation/src/org/osate/aadl2/instantiation/internal/ConnectionArrayExpander.java
@@ -27,7 +27,6 @@ import static org.osate.aadl2.modelsupport.util.AadlUtil.getElementCount;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.function.LongBinaryOperator;
 
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.emf.common.util.EList;
@@ -37,11 +36,9 @@ import org.osate.aadl2.ArrayDimension;
 import org.osate.aadl2.ArraySize;
 import org.osate.aadl2.BasicPropertyAssociation;
 import org.osate.aadl2.Element;
-import org.osate.aadl2.EnumerationLiteral;
 import org.osate.aadl2.Feature;
 import org.osate.aadl2.IntegerLiteral;
 import org.osate.aadl2.ListValue;
-import org.osate.aadl2.NamedValue;
 import org.osate.aadl2.Property;
 import org.osate.aadl2.PropertyAssociation;
 import org.osate.aadl2.PropertyExpression;
@@ -73,6 +70,9 @@ import org.osate.aadl2.util.Aadl2InstanceUtil;
  * connection instances before the expansion runs, see
  * {@link #isStructuralConnectionProperty(Property)}. Both associations are removed from the connection
  * instance they were read from, so they do not survive into the final connection set.
+ * <p>
+ * Which elements a pattern pairs up is {@link ArrayPatternExpansion}; this class decides what a pair of
+ * indices becomes.
  * <p>
  * One expander expands one root once.
  */
@@ -148,7 +148,7 @@ public final class ConnectionArrayExpander {
 				}
 				// only a connection with an array element at one of its ends is expanded
 				if (srcPath.throughArrayElement() || dstPath.throughArrayElement()) {
-					if (new PatternExpansion(conni, false, null, srcPath.sizes(), dstPath.sizes()).expand()) {
+					if (expandPattern(conni, false, null, srcPath.sizes(), dstPath.sizes())) {
 						toRemove.add(conni);
 					}
 				}
@@ -171,8 +171,7 @@ public final class ConnectionArrayExpander {
 						errManager.warning(conni,
 								"Connection pattern specified for connection that does not connect array elements.");
 					} else {
-						if (new PatternExpansion(conni, isOpposite, pattern, srcPath.sizes(), dstPath.sizes())
-								.expand()) {
+						if (expandPattern(conni, isOpposite, pattern, srcPath.sizes(), dstPath.sizes())) {
 							toRemove.add(conni);
 						}
 					}
@@ -273,288 +272,26 @@ public final class ConnectionArrayExpander {
 	}
 
 	/**
-	 * The values of the standard {@code Connection_Pattern} property, which are also the patterns the
-	 * default expansion of an array connection uses. The name of a literal is the name of the constant,
-	 * up to case, which is how {@link #parse(String)} looks a pattern of the model up.
-	 * <p>
-	 * Eleven of the patterns pair up the elements of a source dimension and a destination dimension of
-	 * the same size, and differ only in which source elements take part and which destination element
-	 * each of them is paired with. Those four numbers are the state of such a pattern: the first source
-	 * index, how many elements at the end of the dimension are left out, the step from one source index
-	 * to the next, and the destination index of a source index. The other three patterns have a shape of
-	 * their own and carry none of it.
+	 * Expand a connection instance into the connection instances that its array dimensions, its
+	 * {@code Connection_Pattern} or the default pattern call for.
+	 *
+	 * @param conni the connection instance to expand
+	 * @param isOpposite whether the connection instance goes against the direction of the declaration
+	 * @param patterns the value of {@code Connection_Pattern}, one literal per dimension, or {@code null}
+	 *            to expand with the default pattern
+	 * @param srcSizes the sizes of the array dimensions of the source end, outside in
+	 * @param dstSizes the sizes of the array dimensions of the destination end, outside in
+	 * @return whether the provisional connection instance has to be deleted; a short pattern is rejected
+	 *         without a replacement, while failures that may still describe one scalar connection leave it
+	 *         in place
 	 */
-	private enum ConnectionPattern {
-		/** Every element of the source dimension to every element of the destination dimension. */
-		ALL_TO_ALL,
-
-		/** The one element of the scalar declared source to every element of the destination dimension. */
-		ONE_TO_ALL,
-
-		/** Every element of the source dimension to the one element of the scalar declared destination. */
-		ALL_TO_ONE,
-
-		ONE_TO_ONE(1, 0, 1, (i, size) -> i),
-		NEXT(1, 1, 1, (i, size) -> i + 1),
-		PREVIOUS(2, 0, 1, (i, size) -> i - 1),
-		CYCLIC_NEXT(1, 0, 1, (i, size) -> i == size ? 1 : i + 1),
-		CYCLIC_PREVIOUS(1, 0, 1, (i, size) -> i == 1 ? size : i - 1),
-		NEXT_NEXT(1, 2, 1, (i, size) -> i + 2),
-		PREVIOUS_PREVIOUS(3, 0, 1, (i, size) -> i - 2),
-		CYCLIC_NEXT_NEXT(1, 0, 1, (i, size) -> (i + 1) % size + 1),
-		CYCLIC_PREVIOUS_PREVIOUS(1, 0, 1, (i, size) -> (i + size - 3) % size + 1),
-		EVEN_TO_EVEN(2, 0, 2, (i, size) -> i),
-		ODD_TO_ODD(1, 0, 2, (i, size) -> i);
-
-		/** The first index of the source dimension that takes part. */
-		private final long firstIndex;
-
-		/** How many elements at the end of the source dimension do not take part. */
-		private final int lastIndexOffset;
-
-		/** The distance from one participating source index to the next. */
-		private final long step;
-
-		/** The destination index a source index is paired with, given the size of the dimension. */
-		private final LongBinaryOperator destinationIndex;
-
-		ConnectionPattern() {
-			this(0, 0, 0, null);
-		}
-
-		ConnectionPattern(long firstIndex, int lastIndexOffset, long step, LongBinaryOperator destinationIndex) {
-			this.firstIndex = firstIndex;
-			this.lastIndexOffset = lastIndexOffset;
-			this.step = step;
-			this.destinationIndex = destinationIndex;
-		}
-
-		/**
-		 * The pattern of a {@code Connection_Pattern} enumeration literal.
-		 *
-		 * @param name the name of the literal
-		 * @return the pattern, or {@code null} if this is not a pattern the expansion supports
-		 */
-		static ConnectionPattern parse(String name) {
-			for (ConnectionPattern pattern : values()) {
-				if (pattern.name().equalsIgnoreCase(name)) {
-					return pattern;
-				}
-			}
-			return null;
-		}
-	}
-
-	/**
-	 * One expansion of a connection instance into the connection instances that its array dimensions,
-	 * its {@code Connection_Pattern} or the default pattern call for. The expansion walks the dimensions
-	 * of the two ends outside in, one pattern per dimension, and creates a connection instance for every
-	 * combination of indices the patterns pair up.
-	 * <p>
-	 * The sizes and the patterns do not change; the three cursors and the two index lists are where the
-	 * expansion is. A cursor is advanced by {@link #expandNextDimension(int, int)} only, which also
-	 * restores it, so every level of the recursion sees the cursors and indices it established itself.
-	 * <p>
-	 * Patterns are read in the declarative frame: {@code isOpposite} means that the declared source of
-	 * the connection is the destination of the connection instance, so a pattern that keeps the declared
-	 * source fixed varies the index of the connection instance's destination.
-	 */
-	private final class PatternExpansion {
-		private final ConnectionInstance conni;
-		private final boolean isOpposite;
-
-		/** The pattern per dimension, or {@code null} to expand with the default pattern. */
-		private final List<PropertyExpression> patterns;
-
-		private final List<Integer> srcSizes;
-		private final List<Integer> dstSizes;
-
-		/** The indices collected for the dimensions the expansion has entered, outside in. */
-		private final List<Long> srcIndices = new ArrayList<>();
-		private final List<Long> dstIndices = new ArrayList<>();
-
-		private int offset;
-		private int srcOffset;
-		private int dstOffset;
-
-		/**
-		 * @param conni the connection instance to expand
-		 * @param isOpposite whether the connection instance goes against the direction of the declaration
-		 * @param patterns the value of {@code Connection_Pattern}, one literal per dimension, or
-		 *            {@code null} to expand with the default pattern
-		 * @param srcSizes the sizes of the array dimensions of the source end, outside in
-		 * @param dstSizes the sizes of the array dimensions of the destination end, outside in
-		 */
-		private PatternExpansion(ConnectionInstance conni, boolean isOpposite, List<PropertyExpression> patterns,
-				List<Integer> srcSizes, List<Integer> dstSizes) {
-			this.conni = conni;
-			this.isOpposite = isOpposite;
-			this.patterns = patterns;
-			this.srcSizes = srcSizes;
-			this.dstSizes = dstSizes;
-		}
-
-		/**
-		 * Create the connection instances of the dimension the expansion is at and of every dimension
-		 * inside it.
-		 *
-		 * @return whether the provisional connection instance has to be deleted; a short pattern is rejected
-		 *         without a replacement, while failures that may still describe one scalar connection leave it
-		 *         in place
-		 */
-		private boolean expand() {
-			if (patterns != null && offset == 0
-					&& patterns.size() < Math.max(srcSizes.size(), dstSizes.size())) {
-				errManager.error(conni.getContainingComponentInstance(),
-						"Connection pattern has fewer dimensions than its array ends");
-				return true;
-			}
-			if (patterns != null ? offset >= patterns.size()
-					: srcOffset == srcSizes.size() && dstOffset == dstSizes.size()) {
-				createNewConnection(conni, srcIndices, dstIndices);
-				return true;
-			}
-			String patternName = patternName();
-			ConnectionPattern pattern = ConnectionPattern.parse(patternName);
-			/*
-			 * An unsupported pattern is neither One_To_All nor All_To_One, so it needs an index at both
-			 * ends, and it is reported below, after the size check the supported patterns go through.
-			 */
-			if (!hasIndexFor(End.SOURCE, pattern) || !hasIndexFor(End.DESTINATION, pattern)) {
-				return false;
-			}
-			boolean result = true;
-			if (pattern == ConnectionPattern.ALL_TO_ALL) {
-				for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-					srcIndices.add(i);
-					for (long j = 1; j <= dstSizes.get(dstOffset); j++) {
-						dstIndices.add(j);
-						result &= expandNextDimension(1, 1);
-						dstIndices.removeLast();
-					}
-					srcIndices.removeLast();
-				}
-				return result;
-			}
-			if (pattern == (isOpposite ? ConnectionPattern.ALL_TO_ONE : ConnectionPattern.ONE_TO_ALL)) {
-				// the declared source stays at its one element, so this dimension gives it no index
-				for (long j = 1; j <= dstSizes.get(dstOffset); j++) {
-					dstIndices.add(j);
-					result &= expandNextDimension(0, 1);
-					dstIndices.removeLast();
-				}
-				return result;
-			}
-			if (pattern == (isOpposite ? ConnectionPattern.ONE_TO_ALL : ConnectionPattern.ALL_TO_ONE)) {
-				// the declared destination stays at its one element, so this dimension gives it no index
-				for (long i = 1; i <= srcSizes.get(srcOffset); i++) {
-					srcIndices.add(i);
-					result &= expandNextDimension(1, 0);
-					srcIndices.removeLast();
-				}
-				return result;
-			}
-			// every remaining pattern pairs up two dimensions of the same size
-			int size = srcSizes.get(srcOffset);
-			if (size != dstSizes.get(dstOffset)) {
-				errManager.error(conni,
-						"Array size mismatch (" + patternName + ") on connection " + conni.getFullName() + " in "
-								+ conni.getContainingComponentInstance().getFullName() + ": " + size
-								+ " at source and " + dstSizes.get(dstOffset) + " at destination.");
-				return false;
-			}
-			if (pattern == null) {
-				/*
-				 * A pattern the expansion does not know expands into nothing. Report it and keep the
-				 * connection: returning true would tell the caller that the connection was expanded, and
-				 * the caller would delete it without a replacement.
-				 */
-				errManager.error(conni,
-						"Unsupported connection pattern '" + patternName + "' on connection " + conni.getFullName());
-				return false;
-			}
-			for (long i = pattern.firstIndex; i <= size - pattern.lastIndexOffset; i += pattern.step) {
-				srcIndices.add(i);
-				dstIndices.add(pattern.destinationIndex.applyAsLong(i, size));
-				result &= expandNextDimension(1, 1);
-				dstIndices.removeLast();
-				srcIndices.removeLast();
-			}
-			return result;
-		}
-
-		/**
-		 * Expand the dimensions inside the one the expansion is at with the indices collected for it, then
-		 * undo the step, so that the caller continues with the cursors and indices it had.
-		 *
-		 * @param srcStep one if this dimension took an index from the source end, zero if it did not
-		 * @param dstStep one if this dimension took an index from the destination end, zero if it did not
-		 * @return whether the connection instance was expanded
-		 */
-		private boolean expandNextDimension(int srcStep, int dstStep) {
-			offset++;
-			srcOffset += srcStep;
-			dstOffset += dstStep;
-			boolean result = expand();
-			dstOffset -= dstStep;
-			srcOffset -= srcStep;
-			offset--;
-			return result;
-		}
-
-		/**
-		 * The name of the pattern of the dimension the expansion is at: the name of the enumeration
-		 * literal of {@code Connection_Pattern}, or the name of the pattern the default expansion uses.
-		 * The name is what the array size mismatch message reports, which is why the default expansion
-		 * needs one at all.
-		 */
-		private String patternName() {
-			if (patterns != null) {
-				NamedValue nv = (NamedValue) patterns.get(offset);
-				return ((EnumerationLiteral) nv.getNamedValue()).getName();
-			}
-			/*
-			 * A default pattern pairs dimensions one-to-one. If only one end has dimensions, every element
-			 * on that end maps to the one scalar end. The spelling of these three names is the spelling
-			 * they had.
-			 */
-			if (srcSizes.isEmpty()) {
-				return isOpposite ? "All_to_One" : "One_To_All";
-			}
-			if (dstSizes.isEmpty()) {
-				return isOpposite ? "One_To_All" : "All_to_One";
-			}
-			return "One_to_One";
-		}
-
-		/**
-		 * Does the dimension the expansion is at have an index to give to a declared end of the
-		 * connection? Reports the end having fewer array dimensions than the patterns need indices for.
-		 * <p>
-		 * {@code End} names the declared end here, and the message follows it. Which of the two size
-		 * lists to look in does not: with {@code isOpposite}, the declared source of the connection is the
-		 * destination of the connection instance, so its dimensions are the ones of the destination.
-		 *
-		 * @param end the declared end of the connection
-		 * @param pattern the pattern of the dimension, {@code null} if the model names one that is not
-		 *            supported
-		 * @return whether the pattern has an index for this end
-		 */
-		private boolean hasIndexFor(End end, ConnectionPattern pattern) {
-			// One_To_All takes no index from the declared source, All_To_One none from the declared destination
-			if (pattern == (end == End.SOURCE ? ConnectionPattern.ONE_TO_ALL : ConnectionPattern.ALL_TO_ONE)) {
-				return true;
-			}
-			boolean atInstanceSource = (end == End.SOURCE) != isOpposite;
-			int offsetAtEnd = atInstanceSource ? srcOffset : dstOffset;
-			int dimensions = (atInstanceSource ? srcSizes : dstSizes).size();
-			if (offsetAtEnd < dimensions) {
-				return true;
-			}
-			errManager.error(conni, "Too few indices for connection " + (end == End.SOURCE ? "source" : "destination")
-					+ " for " + conni.getFullName());
-			return false;
-		}
+	private boolean expandPattern(ConnectionInstance conni, boolean isOpposite, List<PropertyExpression> patterns,
+			List<Integer> srcSizes, List<Integer> dstSizes) {
+		var container = conni.getContainingComponentInstance();
+		var subject = new ArrayPatternExpansion.Subject(conni, "connection", conni.getFullName(), container,
+				container.getFullName());
+		return new ArrayPatternExpansion(errManager, subject, isOpposite, patterns, srcSizes, dstSizes,
+				(srcIndices, dstIndices) -> createNewConnection(conni, srcIndices, dstIndices)).expand();
 	}
 
 	private List<Long> getIndices(RecordValue rv, String field) {
